Revert "Merge pull request #1 from sfsu-csc-668-spring-2019/add-renamed-activities-folder"
This reverts commit 6eea06fec6d17cb8b4b71fed1fa36a2e51484791, reversing
changes made to 383750622817aa5a92e166b232a37497b9a95c2c.
