+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Controllers.CreateNewProfile"></activity>
-        <activity android:name=".Controllers.CreateListing" />
-        <activity android:name=".Controllers.MainActivity">
+        <activity android:name=".Activities.CreateListing"></activity>
+        <activity android:name=".Activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
