+++ app/src/main/java/com/example/divvy/Activities/MainActivity.java
@@ -1,15 +1,15 @@
-package com.example.divvy.Controllers;
+package com.example.divvy.Activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.example.divvy.R;
 
-public class CreateNewProfile extends AppCompatActivity {
+public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_create_new_profile);
+        setContentView(R.layout.activity_main);
     }
 }
