--- app/src/main/java/com/example/divvy/Controllers/MainActivity.java
@@ -1,15 +1,0 @@
-package com.example.divvy.Controllers;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-import com.example.divvy.R;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
