+++ app/src/main/res/layout/activity_create_listing.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Controllers.CreateListing">
+    tools:context=".Activities.CreateListing">
 
     <LinearLayout
         android:layout_width="match_parent"
