--- app/src/main/res/layout/activity_create_new_profile.xml
@@ -1,107 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Controllers.CreateNewProfile">
-
-    <TextView
-        android:id="@+id/TextView_title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="60dp"
-        android:text="Join Divvy"
-        android:textSize="36sp"
-        app:layout_constraintBottom_toTopOf="@+id/editText_First_Name"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/textView_First_Name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="First Name"
-        app:layout_constraintBottom_toTopOf="@+id/editText_First_Name"
-        app:layout_constraintStart_toStartOf="@+id/editText_First_Name" />
-
-    <TextView
-        android:id="@+id/textView_Last_Name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Last Name"
-        app:layout_constraintStart_toStartOf="@+id/editText_First_Name"
-        app:layout_constraintTop_toBottomOf="@+id/editText_First_Name" />
-
-    <EditText
-        android:id="@+id/editText_First_Name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:inputType="textPersonName"
-        android:text="Name"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView_First_Name" />
-
-    <EditText
-        android:id="@+id/editText_Last_Name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:inputType="textPersonName"
-        android:text="Name"
-        app:layout_constraintStart_toStartOf="@+id/textView_Last_Name"
-        app:layout_constraintTop_toBottomOf="@+id/textView_Last_Name" />
-
-    <TextView
-        android:id="@+id/textView_Email"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Email Adress"
-        app:layout_constraintBottom_toTopOf="@+id/editText_Email"
-        app:layout_constraintStart_toStartOf="@+id/editText_Last_Name"
-        app:layout_constraintTop_toBottomOf="@+id/editText_Last_Name" />
-
-    <EditText
-        android:id="@+id/editText_Email"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:inputType="textEmailAddress"
-        app:layout_constraintStart_toStartOf="@+id/textView_Email"
-        app:layout_constraintTop_toBottomOf="@+id/textView_Email" />
-
-    <TextView
-        android:id="@+id/textView_Password"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Make A Password"
-        app:layout_constraintStart_toStartOf="@+id/editText_Email"
-        app:layout_constraintTop_toBottomOf="@+id/editText_Email" />
-
-    <EditText
-        android:id="@+id/editText_Password"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:inputType="textPassword"
-        app:layout_constraintStart_toStartOf="@+id/textView_Password"
-        app:layout_constraintTop_toBottomOf="@+id/textView_Password" />
-
-    <Button
-        android:id="@+id/button_Confirm"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="42dp"
-        android:text="Join Now"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/editText_Password" />
-
-</android.support.constraint.ConstraintLayout>
